+++ packages/dashboard-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/dashboard-api
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.5.0
+
 ## 0.5.2
 
 ### Patch Changes