+++ packages/plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/plugin
 
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.5.0
+
 ## 0.7.0
 
 ### Minor Changes