+++ packages/provider-cache/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shardix/provider-cache
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.5.0
+  - @shardix/cache@0.1.4
+
 ## 0.3.3
 
 ### Patch Changes