+++ packages/provider-health/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/provider-health
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.5.0
+
 ## 0.3.3
 
 ### Patch Changes