+++ packages/provider-observability/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/provider-observability
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.5.0
+
 ## 0.4.3
 
 ### Patch Changes