+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shardix/testing
 
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.5.0
+  - @shardix/core@0.5.0
+
 ## 0.7.0
 
 ### Minor Changes